requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-allauth==0.63.3
haikunator==2.1.0
django-soft-delete==0.9.21
kubernetes==28.1.0
PyGithub==2.1.1
jsonschema==4.17.3
python-gitlab==4.1.1
drf-spectacular==0.27.2
dj-rest-auth==6.0.0
boto3==1.34.150
channels[daphne]==4.0.0
channels-redis==4.2.0
daphne==4.1.0
uvicorn[standard]==0.29.0
# Email
django-anymail[mailgun]==10.3
whitenoise==6.7.0
